Move exception to correct destination
