+++ src/Security/LoginFormAuthenticator.php
@@ -5,13 +5,13 @@ declare(strict_types=1);
 namespace Bolt\Security;
 
 use Bolt\Entity\User;
+use Bolt\Exception\DisabledUserLoginAttemptException;
 use Bolt\Repository\UserRepository;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\RouterInterface;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
-use Symfony\Component\Security\Core\Exception\AuthenticationException;
 use Symfony\Component\Security\Core\Exception\InvalidCsrfTokenException;
 use Symfony\Component\Security\Core\Security;
 use Symfony\Component\Security\Core\User\UserInterface;
